setMediaSources update
setMediaSourcesData now checks to make sure the mediaSourceID of the
model exists in the data received. If not, it sets the
currentMediaSource to the first item in the set.

either way, a media source update event is dispatched after
setMediaSourcesData is called. This can be listened to to do things
like bind the ‘you are browsing’ text that exists outside the eureka
component