switching to PureComponent
If you determine a specific component is slow after profiling, you may
change it to inherit from React.PureComponent which implements
shouldComponentUpdate() with a shallow prop and state comparison. If
you are confident you want to write it by hand, you may compare
this.props with nextProps and this.state with nextState and return
false to tell React the update can be skipped.

https://facebook.github.io/react/docs/react-component.html#shouldcompone
ntupdate